server/llm: refactor, generalize name, remove unnecessary try/catches
