+++ src/packages/next/pages/api/v2/openai/embeddings/search.ts
@@ -1,4 +1,4 @@
-import { search } from "@cocalc/server/openai/embeddings-api";
+import { search } from "@cocalc/server/llm/embeddings-api";
 import getAccountId from "lib/account/get-account";
 import getParams from "lib/api/get-params";
