+++ src/packages/server/llm/embeddings-api.ts
