+++ src/packages/server/llm/embeddings.ts
@@ -3,7 +3,7 @@ import getClient from "./client";
 import * as qdrant from "@cocalc/database/qdrant";
 import { getClient as getDB } from "@cocalc/database/pool";
 import checkForAbuse from "./embeddings-abuse";
-import { VertexAIClient } from "./client";
+import { VertexAIClient } from "./vertex-ai-client";
 
 // the vectors we compute using openai's embeddings api get cached for this long
 // in our database since they were last accessed.  Also, this is how long we
